+++ packages/client/src/Ethereum.ts
@@ -27,6 +27,7 @@ export interface EthereumConfig {
     streamRegistryChainRPCs: ChainConnectionInfo
     // most of the above should go into ethereumNetworks configs once ETH-184 is ready
     ethereumNetworks?: Record<string, EthereumNetworkConfig>
+    maxConcurrentContractCalls: number
 }
 
 export const generateEthereumAccount = (): { address: string, privateKey: string } => {