+++ packages/client/src/NetworkNodeFacade.ts
@@ -2,13 +2,14 @@
  * Wrap a network node.
  */
 import { inject, Lifecycle, scoped } from 'tsyringe'
+import EventEmitter from 'eventemitter3'
 import { NetworkNodeOptions, createNetworkNode as _createNetworkNode, MetricsContext } from 'streamr-network'
 import { uuid } from './utils/uuid'
 import { instanceId } from './utils/utils'
 import { pOnce } from './utils/promises'
 import { Context } from './utils/Context'
 import { NetworkConfig, ConfigInjectionToken, TrackerRegistrySmartContract } from './Config'
-import { StreamMessage, StreamPartID, ProxyDirection } from 'streamr-client-protocol'
+import { StreamMessage, StreamPartID, ProxyDirection, } from 'streamr-client-protocol'
 import { DestroySignal } from './DestroySignal'
 import { EthereumConfig, generateEthereumAccount, getMainnetProvider } from './Ethereum'
 import { getTrackerRegistryFromContract } from './registry/getTrackerRegistryFromContract'
@@ -37,7 +38,7 @@ export interface NetworkNodeStub {
     /** @internal */
     stop: () => Promise<unknown>
     /** @internal */
-    openProxyConnection: (streamPartId: StreamPartID, nodeId: string, direction: ProxyDirection) => Promise<void>
+    openProxyConnection: (streamPartId: StreamPartID, nodeId: string, direction: ProxyDirection, userId: string) => Promise<void>
     /** @internal */
     closeProxyConnection: (streamPartId: StreamPartID, nodeId: string, direction: ProxyDirection) => Promise<void>
 }
@@ -45,6 +46,10 @@ export interface NetworkNodeStub {
 export const getEthereumAddressFromNodeId = (nodeId: string): string => {
     const ETHERUM_ADDRESS_LENGTH = 42
     return nodeId.substring(0, ETHERUM_ADDRESS_LENGTH)
+}
+
+export interface Events {
+    start: () => void
 }
 
 /**
@@ -70,6 +75,7 @@ export class NetworkNodeFacade implements Context {
     readonly debug
     private startNodeCalled = false
     private startNodeComplete = false
+    private eventEmitter: EventEmitter<Events>
 
     constructor(
         context: Context,
@@ -83,6 +89,7 @@ export class NetworkNodeFacade implements Context {
         this.ethereumConfig = ethereumConfig
         this.id = instanceId(this)
         this.debug = context.debug.extend(this.id)
+        this.eventEmitter = new EventEmitter<Events>()
         destroySignal.onDestroy.listen(this.destroy)
     }
 
@@ -189,6 +196,8 @@ export class NetworkNodeFacade implements Context {
                 this.debug('stopping node before init >>')
                 await node.stop()
                 this.debug('stopping node before init <<')
+            } else {
+                this.eventEmitter.emit('start')
             }
             this.assertNotDestroyed()
             return node
@@ -240,7 +249,7 @@ export class NetworkNodeFacade implements Context {
             if (this.isStarting()) {
                 await this.startNodeTask()
             }
-            await this.cachedNode!.openProxyConnection(streamPartId, nodeId, direction)
+            await this.cachedNode!.openProxyConnection(streamPartId, nodeId, direction, (await this.authentication.getAddress()))
         } finally {
             this.debug('openProxyConnectionOnStream << %o', streamPartId, nodeId)
         }
@@ -259,6 +268,10 @@ export class NetworkNodeFacade implements Context {
 
     private isStarting(): boolean {
         return !this.cachedNode || !this.startNodeComplete
+    }
+
+    once<E extends keyof Events>(eventName: E, listener: Events[E]): void {
+        this.eventEmitter.once(eventName, listener as any)
     }
 }