+++ packages/client/src/StreamrClient.ts
@@ -30,6 +30,7 @@ import { initContainer } from './Container'
 import { Authentication, AuthenticationInjectionToken } from './Authentication'
 import { StreamStorageRegistry } from './registry/StreamStorageRegistry'
 import { GroupKey } from './encryption/GroupKey'
+import { PublisherKeyExchange } from './encryption/PublisherKeyExchange'
 
 /**
  * @category Important
@@ -76,6 +77,7 @@ export class StreamrClient implements Context {
         this.storageNodeRegistry = container.resolve<StorageNodeRegistry>(StorageNodeRegistry)
         this.streamIdBuilder = container.resolve<StreamIDBuilder>(StreamIDBuilder)
         this.eventEmitter = container.resolve<StreamrClientEventEmitter>(StreamrClientEventEmitter)
+        container.resolve<PublisherKeyExchange>(PublisherKeyExchange) // side effect: activates publisher key exchange
         container.resolve<MetricsPublisher>(MetricsPublisher) // side effect: activates metrics publisher
 
         const context = container.resolve<Context>(Context as any)
@@ -108,9 +110,7 @@ export class StreamrClient implements Context {
         const store = await this.groupKeyStoreFactory.getStore(streamId)
         if (opts.distributionMethod === 'rotate') {
             if (opts.key === undefined) {
-                await store.rotateGroupKey()
-            } else { // eslint-disable-line no-else-return
-                await store.setNextGroupKey(opts.key)
+                await store.rotate(opts.key)
             }
         } else if (opts.distributionMethod === 'rekey') { // eslint-disable-line no-else-return
             await store.rekey(opts.key)
@@ -284,7 +284,7 @@ export class StreamrClient implements Context {
     isStreamPublisher(streamIdOrPath: string, userAddress: EthereumAddress): Promise<boolean> {
         return this.streamRegistry.isStreamPublisher(streamIdOrPath, userAddress)
     }
-    
+
     isStreamSubscriber(streamIdOrPath: string, userAddress: EthereumAddress): Promise<boolean> {
         return this.streamRegistry.isStreamSubscriber(streamIdOrPath, userAddress)
     }
@@ -292,23 +292,23 @@ export class StreamrClient implements Context {
     // --------------------------------------------------------------------------------------------
     // Storage
     // --------------------------------------------------------------------------------------------
-    
+
     addStreamToStorageNode(streamIdOrPath: string, nodeAddress: EthereumAddress): Promise<void> {
         return this.streamStorageRegistry.addStreamToStorageNode(streamIdOrPath, nodeAddress)
     }
-    
+
     removeStreamFromStorageNode(streamIdOrPath: string, nodeAddress: EthereumAddress): Promise<void> {
         return this.streamStorageRegistry.removeStreamFromStorageNode(streamIdOrPath, nodeAddress)
     }
-    
+
     isStoredStream(streamIdOrPath: string, nodeAddress: EthereumAddress): Promise<boolean> {
         return this.streamStorageRegistry.isStoredStream(streamIdOrPath, nodeAddress)
     }
-    
+
     getStoredStreams(nodeAddress: EthereumAddress): Promise<{ streams: Stream[], blockNumber: number }> {
         return this.streamStorageRegistry.getStoredStreams(nodeAddress)
     }
-    
+
     getStorageNodes(streamIdOrPath?: string): Promise<EthereumAddress[]> {
         return this.streamStorageRegistry.getStorageNodes(streamIdOrPath)
     }