+++ packages/client/test/integration/Subscriber.test.ts
@@ -3,7 +3,7 @@ import { Wallet } from '@ethersproject/wallet'
 import { Stream } from '../../src/Stream'
 import { FakeEnvironment } from '../test-utils/fake/FakeEnvironment'
 import StreamrClient, { StreamPermission } from '../../src'
-import { createMockMessage, startPublisherKeyExchangeSubscription } from '../test-utils/utils'
+import { createMockMessage } from '../test-utils/utils'
 import { GroupKey } from '../../src/encryption/GroupKey'
 import { nextValue } from '../../src/utils/iterators'
 import { fastWallet } from 'streamr-test-utils'
@@ -66,7 +66,6 @@ describe('Subscriber', () => {
             }
         })
         await publisher.addEncryptionKey(groupKey, stream.id)
-        await startPublisherKeyExchangeSubscription(publisher)
 
         const sub = await subscriber.subscribe(stream.id)