+++ packages/client/test/integration/parallel-key-exchange.test.ts
@@ -6,7 +6,7 @@ import { StreamPermission } from '../../src/permission'
 import { Stream } from '../../src/Stream'
 import { GroupKey } from '../../src/encryption/GroupKey'
 import { Wallet } from '@ethersproject/wallet'
-import { createMockMessage, startPublisherKeyExchangeSubscription } from '../test-utils/utils'
+import { createMockMessage } from '../test-utils/utils'
 import { MessageRef, StreamMessage } from 'streamr-client-protocol'
 import { wait } from '@streamr/utils'
 import { StreamrClient } from '../../src/StreamrClient'
@@ -45,7 +45,6 @@ describe('parallel key exchange', () => {
                 }
             })
             await publisher.client.addEncryptionKey(publisher.groupKey, stream.id)
-            await startPublisherKeyExchangeSubscription(publisher.client)
         }))
     }, 20000)