+++ packages/client/test/integration/revoke-permissions.test.ts
@@ -73,10 +73,9 @@ describe('revoke permissions', () => {
             auth: {
                 privateKey: subscriberPrivateKey
             },
-            // eslint-disable-next-line no-underscore-dangle
-            _timeouts: {
-                encryptionKeyRequest: 200
-            } as any,
+            decryption: {
+                keyRequestTimeout: 200
+            },
             ...opts
         })
     }